Rename folders and add docstring to test method
