+++ core/domain/email_manager.py
