+++ core/domain/email_manager_test.py
@@ -14,7 +14,7 @@
 
 """Test for sending email to admin."""
 
-from core.domains import email_manager
+from core.domain import email_manager
 from core.tests import app_engine_test_base
 import config
 
@@ -22,6 +22,7 @@ class SendEmailToAdminTest(app_engine_test_base.GenericTestBase):
     """Test for email to be sent to admin email address."""
 
     def test_send_mail_to_admin(self):
+        """Test send_mail_to_admin sends mail to admin email address."""
         user_email = 'user1@example.com'
 
         messages = self.mail_stub.get_sent_messages(
